refactor: Name 객체 생성
